add functions into mapService
