+++ src/components/mermaidMap/mapService.js
@@ -1,3 +1,6 @@
+import React from 'react'
+import { createRoot } from 'react-dom/client'
+import Popup from './Popup'
 import maplibregl from 'maplibre-gl'
 import mapPin from '../../assets/map-pin.png'
 
@@ -352,5 +355,96 @@ export const handleMapOnWheel = (mapCurrent, handleZoomDisplayHelpText) => {
         handleZoomDisplayHelpText(false)
       }, 1500)
     }
+  })
+}
+
+export const addClusterSourceAndLayers = (map, data) => {
+  map.addSource('mapMarkers', {
+    type: 'geojson',
+    data: {
+      type: 'FeatureCollection',
+      features: data.map((site) => ({
+        type: 'Feature',
+        geometry: {
+          type: 'Point',
+          coordinates: [site.longitude, site.latitude],
+        },
+        properties: site,
+      })),
+    },
+    cluster: true,
+    clusterMaxZoom: 14,
+    clusterRadius: 50,
+  })
+
+  map.addLayer({
+    id: 'clusters',
+    type: 'circle',
+    source: 'mapMarkers',
+    filter: ['has', 'point_count'],
+    paint: {
+      'circle-color': ['step', ['get', 'point_count'], '#51bbd6', 100, '#f1f075', 750, '#f28cb1'],
+      'circle-radius': ['step', ['get', 'point_count'], 20, 100, 30, 750, 40],
+    },
+  })
+
+  map.addLayer({
+    id: 'cluster-count',
+    type: 'symbol',
+    source: 'mapMarkers',
+    filter: ['has', 'point_count'],
+    layout: {
+      'text-field': ['get', 'point_count_abbreviated'],
+      'text-font': ['Open Sans Regular', 'Arial Unicode MS Regular'],
+      'text-size': 12,
+    },
+  })
+
+  map.addLayer({
+    id: 'unclustered-point',
+    type: 'circle',
+    source: 'mapMarkers',
+    filter: ['!', ['has', 'point_count']],
+    paint: {
+      'circle-color': '#11b4da',
+      'circle-radius': 4,
+      'circle-stroke-width': 1,
+      'circle-stroke-color': '#fff',
+    },
+  })
+}
+
+export const addClusterEventListeners = (map, popUpRef, choices) => {
+  map.on('click', 'clusters', ({ features }) => {
+    const clusterId = features[0].properties.cluster_id
+
+    map.getSource('mapMarkers').getClusterExpansionZoom(clusterId, (err, zoom) => {
+      if (err) {
+        return
+      }
+
+      map.easeTo({
+        center: features[0].geometry.coordinates,
+        zoom: zoom,
+      })
+    })
+  })
+
+  map.on('click', 'unclustered-point', (e) => {
+    const coordinates = e.features[0].geometry.coordinates.slice()
+    const markerProperty = e.features[0].properties
+    const popupNode = document.createElement('div')
+    const reactRoot = createRoot(popupNode)
+
+    reactRoot.render(<Popup properties={markerProperty} choices={choices} />)
+
+    popUpRef.current.setLngLat(coordinates).setDOMContent(popupNode).addTo(map)
+  })
+
+  map.on('mouseenter', 'clusters', () => {
+    map.getCanvas().style.cursor = 'pointer'
+  })
+  map.on('mouseleave', 'clusters', () => {
+    map.getCanvas().style.cursor = ''
   })
 }
